[고양이 비만도 계산기] ui수정 및 계산 로직 수정완료

+++ lib/core/router/router.dart
@@ -9,14 +9,15 @@ final router = GoRouter(
       builder: (context, state) => const MainScreen(),
     ),
     GoRoute(
-        path: '/main/result',
-        builder: (context, state) {
-          double weight = double.parse(state.queryParameters['weight']!);
-          double bcs = double.parse(state.queryParameters['bcs']!);
-          return LoadingScreen(
-            weight: weight,
-            bcs: bcs,
-          );
-        })
+      path: '/main/result',
+      builder: (context, state) {
+        double weight = double.parse(state.queryParameters['weight']!);
+        int bcs = int.parse(state.queryParameters['bcs']!);
+        return LoadingScreen(
+          weight: weight,
+          bcs: bcs,
+        );
+      },
+    ),
   ],
 );
